oh the joys of floating-point arithmetic. possible fix for CylCutter::edgeDrop which was causing Issue-7
an edge is defined by its endpoints p1 and p2
This has to do with computing cc-points on edges, where the coordinates 
are very close to eachoter.
In the bug-report there were triangles with "small" abs(p1.x - p2.x) which 
means we must use the y-coordinate to compute the height of the cc-point.