look for git "by hand" if Cmake can't find it
version_string.cmake uses the FindGit.cmake module to look for git,
but older versions of Cmake don't have this module.  Specifically,
cmake 2.8.1 in Lucid doesn't have it.

This commit adds a "manual" check for the git executable if the normal
FindGit.cmake way can't be run.  This makes opencamlib build on Lucid.