build the ocl python module specifically for python2
Without this commit, cmake will pick the highest available version number
of PythonLibs to build against.  On machines with both python2 and python3
dev packages installed, this would incorrectly try to build for python3.

With this commit, cmake will pick the highest available 2.x.y version
of PythonLibs, which is what we want.