Protect against race in CLPoint::liftZ
Since moving to a machine with 12 hw threads, I have seen occasional
segfaults from a double free in CLPoint::liftZ. This is consistent with
several threads calling liftZ on the same CLPoint instance, racing to
delete the previous instance.

To avoid the race, this diff wraps the raw pointer in a std::atomic, and
then uses std::atomic<T>::compare_exchange_strong for an atomic
replacement without double frees or leaks.

Another option would be to use a mutex.